+++ korge/src/commonMain/kotlin/korlibs/korge/view/filter/BlurFilter.kt
@@ -5,7 +5,7 @@ import korlibs.math.geom.*
 import kotlin.math.*
 
 class BlurFilter(
-    radius: Double = 4.0,
+    radius: Float = 4f,
     expandBorder: Boolean = true,
     @ViewProperty
     var optimize: Boolean = true
@@ -20,13 +20,13 @@ class BlurFilter(
             vertical.expandBorder = value
         }
     @ViewProperty
-    var radius: Double = radius
+    var radius: Float = radius
         set(value) {
             field = value
             horizontal.radius = radius
             vertical.radius = radius
         }
-    override val recommendedFilterScale: Double get() = if (!optimize || radius <= 2.0) 1.0 else 1.0 / log2(radius * 0.5)
+    override val recommendedFilterScale: Float get() = if (!optimize || radius <= 2.0) 1f else 1f / log2(radius.toFloat() * 0.5f)
 
-    override val isIdentity: Boolean get() = radius == 0.0
+    override val isIdentity: Boolean get() = radius == 0f
 }
